requirements.txt
Django==2.2
django-filter==21.1
djangorestframework==3.11
pytz==2021.3
sqlparse==0.4.2
drf-yasg==1.17.1
